+++ .github/workflows/pull_request.yaml
@@ -27,6 +27,6 @@ jobs:
       config_file: '.github/.licenserc.yaml'
       source_dir: ''
       compilers: '["gcc-11"]'
-      doc_target: 'fastscf_cxx_api'
+      doc_target: 'scf_cxx_api'
       build_fail_on_warning: false
     secrets: inherit
